requirements.txt
flask
torch
transformers
sentence-transformers
pandas
pyarrow
python-dotenv
requests
gunicorn
spacy
tqdm
langchain
sources
numpy
pytest
pytest-mock
ragas
datasets
langchain-community
langchain-huggingface
pyahocorasick
openai
selenium
webdriver-manager
pytchat